# src/main/java/ejada/task/taskManagement/controller/dtos/CreatedTaskManagementResponseModel.java
package ejada.task.taskManagement.controller.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class CreatedTaskManagementResponseModel {

    private Long id;
}