# swaggers.json
{
  "openapi":"3.1.0",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8081",
      "description":"Generated server url"
    }
  ],
  "paths":{
    "/task-management":{
      "get":{
        "tags":[
          "task-management-controller"
        ],
        "operationId":"findAllTaskManagements",
        "parameters":[
          {
            "name":"taskManagementFilter",
            "in":"query",
            "required":true,
            "schema":{
              "$ref":"#/components/schemas/TaskManagementFilter"
            }
          },
          {
            "name":"pageable",
            "in":"query",
            "required":true,
            "schema":{
              "$ref":"#/components/schemas/Pageable"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/PageTaskManagementDetailsResponseModel"
                }
              }
            }
          }
        }
      },
      "post":{
        "tags":[
          "task-management-controller"
        ],
        "operationId":"createTaskManagement",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/CreateOrUpdateTaskManagementRequestDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/CreatedTaskManagementResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/register":{
      "post":{
        "tags":[
          "auth-controller"
        ],
        "operationId":"registerUser",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/User"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK"
          }
        }
      }
    },
    "/authenticate":{
      "post":{
        "tags":[
          "auth-controller"
        ],
        "operationId":"createAuthenticationToken",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/AuthRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"string"
                }
              }
            }
          }
        }
      }
    },
    "/task-management/{taskManagementId}":{
      "delete":{
        "tags":[
          "task-management-controller"
        ],
        "operationId":"deleteTaskManagementById",
        "parameters":[
          {
            "name":"taskManagementId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK"
          }
        }
      },
      "patch":{
        "tags":[
          "task-management-controller"
        ],
        "operationId":"updateTaskManagementById",
        "parameters":[
          {
            "name":"taskManagementId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/CreateOrUpdateTaskManagementRequestDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK"
          }
        }
      }
    },
    "/task-management/{taskManagementId}/details":{
      "get":{
        "tags":[
          "task-management-controller"
        ],
        "operationId":"findTaskManagementById",
        "parameters":[
          {
            "name":"taskManagementId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/TaskManagementDetailsResponseModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "CreateOrUpdateTaskManagementRequestDTO":{
        "type":"object",
        "properties":{
          "title":{
            "type":"string"
          },
          "description":{
            "type":"string"
          },
          "status":{
            "type":"string",
            "enum":[
              "TO_DO",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "priority":{
            "type":"string",
            "enum":[
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "createdBy":{
            "type":"string"
          },
          "dueDate":{
            "type":"string"
          }
        }
      },
      "CreatedTaskManagementResponseModel":{
        "type":"object",
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          }
        }
      },
      "User":{
        "type":"object",
        "properties":{
          "name":{
            "type":"string"
          },
          "username":{
            "type":"string"
          },
          "password":{
            "type":"string"
          }
        }
      },
      "AuthRequest":{
        "type":"object",
        "properties":{
          "username":{
            "type":"string"
          },
          "password":{
            "type":"string"
          }
        }
      },
      "TaskManagementFilter":{
        "type":"object",
        "properties":{
          "status":{
            "type":"string",
            "enum":[
              "TO_DO",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "priority":{
            "type":"string",
            "enum":[
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          }
        }
      },
      "Pageable":{
        "type":"object",
        "properties":{
          "page":{
            "type":"integer",
            "format":"int32",
            "minimum":0
          },
          "size":{
            "type":"integer",
            "format":"int32",
            "minimum":1
          },
          "sort":{
            "type":"array",
            "items":{
              "type":"string"
            }
          }
        }
      },
      "PageTaskManagementDetailsResponseModel":{
        "type":"object",
        "properties":{
          "totalElements":{
            "type":"integer",
            "format":"int64"
          },
          "totalPages":{
            "type":"integer",
            "format":"int32"
          },
          "pageable":{
            "$ref":"#/components/schemas/PageableObject"
          },
          "first":{
            "type":"boolean"
          },
          "last":{
            "type":"boolean"
          },
          "size":{
            "type":"integer",
            "format":"int32"
          },
          "content":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/TaskManagementDetailsResponseModel"
            }
          },
          "number":{
            "type":"integer",
            "format":"int32"
          },
          "sort":{
            "$ref":"#/components/schemas/SortObject"
          },
          "numberOfElements":{
            "type":"integer",
            "format":"int32"
          },
          "empty":{
            "type":"boolean"
          }
        }
      },
      "PageableObject":{
        "type":"object",
        "properties":{
          "paged":{
            "type":"boolean"
          },
          "pageNumber":{
            "type":"integer",
            "format":"int32"
          },
          "pageSize":{
            "type":"integer",
            "format":"int32"
          },
          "offset":{
            "type":"integer",
            "format":"int64"
          },
          "sort":{
            "$ref":"#/components/schemas/SortObject"
          },
          "unpaged":{
            "type":"boolean"
          }
        }
      },
      "SortObject":{
        "type":"object",
        "properties":{
          "sorted":{
            "type":"boolean"
          },
          "empty":{
            "type":"boolean"
          },
          "unsorted":{
            "type":"boolean"
          }
        }
      },
      "TaskManagementDetailsResponseModel":{
        "type":"object",
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          },
          "title":{
            "type":"string"
          },
          "description":{
            "type":"string"
          },
          "status":{
            "type":"string"
          },
          "priority":{
            "type":"string"
          },
          "createdBy":{
            "type":"string"
          },
          "dueDate":{
            "type":"string"
          }
        }
      }
    }
  }
}